CCP-461: corrected extension name in package file, and reorganized components
