+++ frontend/components/PriceStriked/index.jsx
@@ -1,48 +1,99 @@
-import React, { Fragment } from 'react';
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import I18n from '@shopgate/pwa-common/components/I18n';
-import PlaceholderLabel from '../PlaceholderLabel/index';
-import StrikePrice from './components/PriceStriked/index';
 import styles from './style';
 
 /**
- * The PriceStriked component.
- * @param {Object} props The component props.
+ * Calculates the angle for the strike-through line
+ * @param {HTMLElement} element The price element
+ * @returns {number} the calculated angle
+ */
+const calcAngle = (element) => {
+  /**
+   * If the element is in a hidden container the browser won't calculate it's size.
+   * In that case we clone in into a visible container and then remove it again.
+   */
+  const cloned = element.cloneNode(true);
+  document.body.appendChild(cloned);
+
+  const width = cloned.offsetWidth;
+  const height = cloned.offsetHeight;
+
+  document.body.removeChild(cloned);
+
+  // Calculate the correct angle for the strike-through line
+  return Math.round(90 - (Math.atan(width / height) * (180 / Math.PI)));
+};
+
+/**
+ * The price striked component
+ * @param {Object} props The component props
+ * @param {number} props.value The sriked price of the product
+ * @param {string} props.currency The currency of the price
+ * @param {string} [props.className] CSS classes
  * @return {JSX}
  */
-const PriceStriked = ({ price }) => (
-  <PlaceholderLabel
-    className={styles.placeholder}
-    ready={(price !== null)}
-  >
-    {(price && price.msrp > 0 && price.unitPrice !== price.msrp) && (
-      <Fragment>
-        <I18n.Text
-          string="price.msrp"
-          className={styles.msrp}
-        />
-        <StrikePrice
-          className={styles.msrpStriked}
-          value={price.msrp}
-          currency={price.currency}
-        />
-      </Fragment>
-    )}
-    {(price && !price.msrp && price.unitPriceStriked > 0) && (
-      <StrikePrice
-        value={price.unitPriceStriked}
-        currency={price.currency}
-      />
-    )}
-  </PlaceholderLabel>
-);
+class PriceStriked extends Component {
+  static propTypes = {
+    currency: PropTypes.string.isRequired,
+    value: PropTypes.number.isRequired,
+    className: PropTypes.string,
+  };
 
-PriceStriked.propTypes = {
-  price: PropTypes.shape(),
-};
+  static defaultProps = {
+    className: '',
+  };
 
-PriceStriked.defaultProps = {
-  price: null,
-};
+  /**
+   * Constructor
+   * @param {Object} props The component props
+   */
+  constructor(props) {
+    super(props);
+    this.angle = null;
+    this.element = null;
+  }
+
+  /**
+   * Updates the component one more time with the calculated angle
+   * based on the DOM element.
+   */
+  componentDidMount() {
+    if (this.setAngle()) {
+      this.forceUpdate();
+    }
+  }
+
+  /**
+   * Sets the calculated angle for the DOM element
+   * and returns true if succeeded.
+   * @param {Object} element The target element.
+   * @returns {boolean}
+   */
+  setAngle = () => {
+    if (this.element) {
+      this.angle = calcAngle(this.element);
+      return true;
+    }
+
+    return false;
+  };
+
+  /**
+   * Renders the component.
+   * @returns {JSX}
+   */
+  render() {
+    const angleStyle = this.angle ? styles.getAngleStyle(this.angle) : '';
+
+    return (
+      <div className={`${styles.basic} ${this.props.className} ${angleStyle}`}>
+        <span ref={(ref) => { this.element = ref; }}>
+          <I18n.Price price={this.props.value} currency={this.props.currency} />
+        </span>
+      </div>
+    );
+  }
+}
 
 export default PriceStriked;
