+++ frontend/components/PriceStriked/style.js
@@ -1,26 +1,37 @@
 import { css } from 'glamor';
 import { themeConfig } from '@shopgate/pwa-common/helpers/config';
 
-const placeholder = css({
-  height: 16,
-  width: '70%',
-  marginTop: 5,
-  marginBottom: 2,
-}).toString();
-
-const msrp = css({
+const basic = css({
   color: themeConfig.colors.shade3,
-  fontSize: '0.875rem',
-  marginRight: themeConfig.variables.gap.small / 2,
+  whiteSpace: 'nowrap',
+  '& span': {
+    position: 'relative',
+    ':before': { // The strike-through effect is achieved by using the :before pseudo-class.
+      borderColor: 'currentColor',
+      content: '""',
+      position: 'absolute',
+      left: 0,
+      right: 0,
+      top: '50%',
+    },
+  },
 }).toString();
 
-const msrpStriked = css({
-  display: 'inline',
-  fontSize: '0.875rem',
+/**
+ * Returns a class for the rotated strike-through line.
+ * @param {number} angle The calculated angle for the strike-through line.
+ * @returns {string} Classname
+ */
+const getAngleStyle = angle => css({
+  '& span': {
+    ':before': {
+      borderTop: '1px solid',
+      transform: `rotate(-${angle}deg)`,
+    },
+  },
 }).toString();
 
 export default {
-  placeholder,
-  msrp,
-  msrpStriked,
+  basic,
+  getAngleStyle,
 };
