+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@shopgate-project/city-gear_30059",
+  "name": "@shopgate/minimum-advertising-pricing",
   "version": "0.2.3",
   "description": "Shopgate Connect City Gear custom adaptions.",
   "license": "Apache-2.0",
